Moved non-GitRepoInspector exceptions to their own files.
