+++ src/Verlite.Core/GitRepoInspector.cs
@@ -1,45 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace Verlite
 {
-	/// <summary>
-	/// An exception thrown when the repository could not be deepened.
-	/// </summary>
-	/// <seealso cref="RepoInspectionException"/>
-	[ExcludeFromCodeCoverage]
-	public class AutoDeepenException : RepoInspectionException
-	{
-		/// <summary>
-		/// Initializes a new instance of <see cref="AutoDeepenException"/> class.
-		/// </summary>
-		public AutoDeepenException() : base("Failed to automatically deepen the repository") { }
-		/// <summary>
-		/// Initializes a new instance of the <see cref="AutoDeepenException"/> class.
-		/// </summary>
-		/// <param name="message">The message that describes the error.</param>
-		public AutoDeepenException(string message) : base($"Failed to automatically deepen the repository: {message}") { }
-		/// <summary>
-		/// Initializes a new instance of the <see cref="AutoDeepenException"/> class.
-		/// </summary>
-		/// <param name="parent">An inner exception.</param>
-		internal AutoDeepenException(CommandException parent) : base("Failed to automatically deepen the repository: " + parent.Message, parent) { }
-	}
-	/// <summary>
-	/// An exception thrown when an operation could not be completed due to being too shallow.
-	/// </summary>
-	/// <seealso cref="RepoInspectionException"/>
-	public class RepoTooShallowException : RepoInspectionException
-	{
-		/// <summary>
-		/// Initializes a new instance of the <see cref="RepoTooShallowException"/> class.
-		/// </summary>
-		internal RepoTooShallowException() : base("No version tag found before shallow clone reached end.") { }
-	}
 	/// <summary>
 	/// An exception that gets thrown when trying to initialize a <see cref="GitRepoInspector"/> from a directory not contianing a git repository.
 	/// </summary>
